Added Support for Global Style options
New static class GStyle contains the attributes for axes and datasets.
These style options are set as the default for their classes when they
are created. (Changing GStyle after an H1F has been created will have
no effect. The style options are copied when the H1F is created)